asr now runs stem-to-stern with externel generax hack
