+++ topiary/topiary/__init__.py
@@ -8,3 +8,4 @@ from .opentree import get_ott_id, get_species_tree
 from . import ncbi
 from . import util
 from . import raxml
+from . import reconcile
